Move shared mock creation logic to test helpers file
